Support server command (#4389)
* server command

* Update

* Fix build error

* Fix incorrect return value

* Refactor

* Fix ipc error

* Refactor pipe packet buffer

* Optimize name

* Refactor ProcessFactory

* Remove legacy code

* Optimize

* Optimize code

* Refactor, add message_bus.cc

* Optimize

* Fix

* Fix tests

* Optimize code

* Optimize code[2]

* Fix core-tests

* Fix tests[2]

* Update example

* Fix tests

* Fix tests[2]

* Optimize code, fix core-tests

* Optimize code[3], add tests

* Optimize code[4], add command supports for task_worker

* revert name

* Optimize performance

+++ ext-src/swoole_coroutine_scheduler.cc
@@ -123,7 +123,7 @@ void php_swoole_coroutine_scheduler_minit(int module_number) {
 static zend_fcall_info_cache exit_condition_fci_cache;
 static bool exit_condition_cleaner;
 
-static bool php_swoole_coroutine_reactor_can_exit(Reactor *reactor, int &event_num) {
+static bool php_swoole_coroutine_reactor_can_exit(Reactor *reactor, size_t &event_num) {
     zval retval;
     int success;
